handle new col

+++ database/migrations/2025_04_18_073801_update_courses_table_columns.php
@@ -12,20 +12,24 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('courses', function (Blueprint $table) {
-            //
+            if (! Schema::hasColumn('courses', 'daysInWeek')) {
+                $table->string('daysInWeek')->nullable()->after('hours');
+            }
         });
-        Schema::create('course_instructor', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('course_id')->constrained()->onDelete('cascade');
-            $table->foreignId('instructor_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
-        Schema::create('certificate_course', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('course_id')->constrained()->onDelete('cascade');
-            $table->foreignId('certificate_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        // Schema::create('course_instructor', function (Blueprint $table) {
+        //     $table->id();
+        //     $table->foreignId('course_id')->constrained()->onDelete('cascade');
+        //     $table->foreignId('instructor_id')->constrained()->onDelete('cascade');
+        //     $table->timestamps();
+        // });
+        if (! Schema::hasTable('certificate_course')) {
+            Schema::create('certificate_course', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('course_id')->constrained()->onDelete('cascade');
+                $table->foreignId('certificate_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
